Minor refactor of targetted frontend source files for test
Broke testable functions out from vrms-events and current-projects into their associated -utils files.
Added includeing the current-projects-utils.js script in current-projects
Updated ESM-syntax vrms-events to use the extension .mjs, so it will be interpreted correctly by the test runner.
Updated files that call vrms events to use the new extension.
