+++ _includes/current-projects.html
@@ -76,5 +76,6 @@
   </div>
 </section>
 
+<script src="/assets/js/current-projects-utils.js"></script>
 <script src="/assets/js/current-projects.js"></script>
 <script src="../assets/js/elements/noResultsMessageFilter.js"></script>
